Move initial document state to JSON file
