+++ client/src/renderer/components/App/demo.json
