+++ client/src/renderer/components/Editor/Elements/Heading.jsx
@@ -3,7 +3,7 @@
 import { h } from "preact";
 import { renderNodeList } from "dom/render";
 
-export default function Heading(props) {
+export function Heading(props) {
   // Ensure level is valid
   if (props.level < 1) {
     throw new RenderError(`Invalid heading level '${props.level}'`);
