+++ client/src/renderer/components/Editor/Elements/Paragraph.jsx
@@ -1,7 +1,9 @@
+/** @format */
+
 import { h } from "preact";
 import { renderNodeList } from "dom/render";
 
-export default (props) => (
+export const Paragraph = props => (
   <p class={props.class && `__editor__${props.class}`}>
     {renderNodeList(props.children)}
   </p>
