+++ client/src/renderer/components/Editor/Elements/Span.jsx
@@ -1,0 +1,8 @@
+/** @format */
+
+import { h } from "preact";
+import { renderNodeList } from "dom/render";
+
+export function Span(props) {
+  return <span>{renderNodeList(props.children)}</span>;
+}
