+++ client/src/renderer/components/Editor/Elements/index.jsx
@@ -1,0 +1,5 @@
+/** @format */
+
+export { Heading } from "./Heading";
+export { Paragraph } from "./Paragraph";
+export { Span } from "./Span";
