+++ client/src/renderer/components/Editor/Nodes/index.jsx
@@ -1,0 +1,7 @@
+/** @format */
+
+import { h } from "preact";
+
+export const LineBreak = () => <br />;
+export const PageBreak = () => <p>----- Page break -----</p>;
+export const Text = props => props.content;
