+++ client/src/renderer/index.jsx
@@ -5,7 +5,7 @@ import "styles/application.scss";
 import { h, render } from "preact";
 import { Provider } from "react-redux";
 
-import Application from "components/Application";
+import Application from "components/App";
 import store from "redux/store";
 
 render(
